+++ frontend/src/components/PresetList.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import React, { ReactNode } from "react"
+import type { ReactNode, JSX } from "react"
 
 import Link from "next/link"
 
@@ -10,35 +10,34 @@ import AsyncButton from "@/components/AsyncButton"
 import Button from "@/components/Button"
 import LoadingSpinner from "@/components/loading.svg"
 import { PlatformIcon } from "@/components/PlatformSelect/PlatformIcon"
-import styles from "@/components/ScratchList.module.scss"
-import * as api from "@/lib/api"
+import { type Preset, usePaginated } from "@/lib/api"
 import { presetUrl } from "@/lib/api/urls"
 import useTranslation from "@/lib/i18n/translate"
 
 export interface Props {
     url?: string
     className?: string
-    item?: ({ preset }: { preset: api.Preset }) => React.JSX.Element
+    item?: ({ preset }: { preset: Preset }) => JSX.Element
     emptyButtonLabel?: ReactNode
 }
 
-export function PresetList({ url, className, item, emptyButtonLabel }: Props): React.JSX.Element {
-    const { results, isLoading, hasNext, loadNext } = api.usePaginated<api.Preset>(url || "/preset")
+export function PresetList({ url, className, item, emptyButtonLabel }: Props): JSX.Element {
+    const { results, isLoading, hasNext, loadNext } = usePaginated<Preset>(url || "/preset")
     if (results.length === 0 && isLoading) {
-        return <div className={classNames(styles.loading, className)}>
+        return <div className={classNames("flex justify-center items-center gap-[0.5em] p-[1em] opacity-50", className)}>
             <LoadingSpinner width="1.5em" height="1.5em" />
             Just a moment...
         </div>
     }
 
-    const Item = item || PresetItem
+    const Item = item ?? PresetItem
 
     return (
-        <ul className={classNames(styles.list, "rounded-md border-gray-6 text-sm", className)}>
+        <ul className={classNames("flex flex-col justify-center gap-[0.5em] overflow-hidden p-[1em] border-[1px] border-[solid] rounded-md border-gray-6 text-sm", className)}>
             {results.map(preset => (
                 <Item hideIcon key={preset.id} preset={preset} />
             ))}
-            {results.length === 0 && emptyButtonLabel && <li className={styles.button}>
+            {results.length === 0 && emptyButtonLabel && <li className={"col-[span_var(--num-columns,_1)] mt-[0.5em] flex items-center justify-center opacity-70"}>
                 <Link href="/new">
 
                     <Button>
@@ -47,7 +46,7 @@ export function PresetList({ url, className, item, emptyButtonLabel }: Props): R
 
                 </Link>
             </li>}
-            {hasNext && <li className={styles.button}>
+            {hasNext && <li className={"col-[span_var(--num-columns,_1)] mt-[0.5em] flex items-center justify-center opacity-70"}>
                 <AsyncButton onClick={loadNext}>
                     Show more
                 </AsyncButton>
@@ -56,16 +55,14 @@ export function PresetList({ url, className, item, emptyButtonLabel }: Props): R
     )
 }
 
-export function PresetItem({ preset, hideIcon }: { preset: api.Preset, hideIcon?: boolean }): React.JSX.Element {
+export function PresetItem({ preset, hideIcon }: { preset: Preset, hideIcon?: boolean }): JSX.Element {
     const compilersTranslation = useTranslation("compilers")
     const compilerName = compilersTranslation.t(preset.compiler)
 
     return (
         <div className="rounded-md border border-gray-6 p-[1em] text-sm">
             <div className="flex items-center gap-2">
-                {
-                    hideIcon ? <></>:<PlatformIcon platform={preset.platform} className="w-[1.2em]"/>
-                }
+                {hideIcon && <PlatformIcon platform={preset.platform} className="w-[1.2em]"/>}
                 <a className="font-semibold hover:text-[var(--link)]" href={presetUrl(preset)}>
                     {preset.name}
                 </a>
